Complete basic calculator exercises

+++ exercises/02_basic_calculator/01_integers/src/lib.rs
@@ -1,7 +1,7 @@
 fn compute(a: u32, b: u32) -> u32 {
     // TODO: change the line below to fix the compiler error and make the tests pass.
     // TODO: 修改下面这行以修复编译器错误并使测试通过。
-    let multiplier: u8 = 4;
+    let multiplier: u32 = 4;
     a + b * multiplier
 }
